Add pdf_page.py which has the PdfPage class
The new PdfPage class is created to represent a PDF page. It includes a method to extract the page number. This class is intended to be a base class to be inherited by other classes for specific pages. This would help in better organization of code and improve the readability as well as the reusability of it.